[ELY-713] Pull out additional utility methods for creating password hashes.
